pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Cherepanova Xenia <cherepanova.xenia@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
tortoise = "^0.1.1"
tortoise-orm = "^0.20.0"
pydantic = "^2.5.2"
email-validator = "^2.1.0.post1"
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
python-multipart = "^0.0.6"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
aiogram = "^2.14.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
